+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 
 :warning: This server code is versioned separately to the download of the [Hiro game framework](https://heroiclabs.com/hiro/). :warning:
 
+## [Unreleased]
+### Added
+- Tutorials can be marked as skipped, with a new "Skip" function and "tutorialSkipped" event.
+
 ## [1.35.0] - 2026-07-29
 ### Added
 - Streaks have new options "grace_window_sec" and "auto_claim" for configuration.